requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sentence-transformers==2.7.0
transformers==4.50.0
torch~=2.7.1
openai-whisper>=20230314
Pillow>=10.0.0,<11.0.0
pydantic~=2.11.7
numpy>=1.24.0,<2.0.0
ffmpeg-python>=0.2.0  # for audio loading
whisper~=1.1.10
hf_xet
injector~=0.22.0
boto3>=1.30.0,<2.0.0
aiofiles>=23.1.0,<24.0.0
httpx>=0.27.0,<1.0.0
s3fs>=2023.6.0,<2024.0.0
librosa>=0.10.0,<0.11.0
soundfile>=0.12.1,<0.13.0
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.0,<1.0.0
slowapi
diskcache>=5.6.1,<6.0.0
qdrant-client>=1.0.0,<2.0.0

